Afternote DTO 구조 개선 및 패키지 정리
- 여러 파일에 분산되어 있던 `MusicSearchResponse`, `MusicTrack`, `AfternotePlaylist` 등 음악 관련 DTO들을 `MusicDto.kt`로 통합
- `request` 및 `response` 패키지로 분리되어 있던 DTO들을 상위 `dto` 패키지로 이동 및 통합
- 패키지 구조 변경에 따른 `Repository`, `Service`, `Mapper` 클래스의 import 문 수정
