--- feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/dto/AfternoteCredentials.kt
@@ -1,10 +1,0 @@
-package com.afternote.feature.afternote.data.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AfternoteCredentials(
-    @SerialName("id") val id: String? = null,
-    @SerialName("password") val password: String? = null,
-)
