--- feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/dto/AfternoteMemorialVideo.kt
@@ -1,10 +1,0 @@
-package com.afternote.feature.afternote.data.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AfternoteMemorialVideo(
-    @SerialName("videoUrl") val videoUrl: String? = null,
-    @SerialName("thumbnailUrl") val thumbnailUrl: String? = null,
-)
