--- feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/dto/AfternoteReceiverRef.kt
@@ -1,12 +1,0 @@
-package com.afternote.feature.afternote.data.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-/**
- * 수신자 참조 (API ReceiverRequest). 수신자 목록은 모든 카테고리에서 사용 가능.
- */
-@Serializable
-data class AfternoteReceiverRef(
-    @SerialName("receiverId") val receiverId: Long,
-)
