--- feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/dto/MusicTrack.kt
@@ -1,14 +1,0 @@
-package com.afternote.feature.afternote.data.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-/**
- * Single track in music search response.
- */
-@Serializable
-data class MusicTrack(
-    @SerialName("artist") val artist: String,
-    @SerialName("title") val title: String,
-    @SerialName("albumImageUrl") val albumImageUrl: String? = null,
-)
